🐛 Use dummy user if papertrail whodunnit is nil
We've recently seen some Sentry errors trying to call `.split` on a
`nil` object. After looking closer at the issue, it seems we're
occasionally having trouble generating our list of `AuditEvent` records
due to some `PaperTrail::Version` objects having a `nil` value for
`whodunnit`.

This commit avoids the runtime error by moving our `dummy_user` helper
method out of `ApplicationController` and into a reusable `audit_helper`
module. This allows us to return a dummy user record whenever the
`PaperTrail` `whodunnit` value is `nil`.
